Generalize skip function
Make it take a more general parameter and place it in utils.
